requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
tqdm>=4.62.0
pyyaml>=6.0
einops>=0.6.0
h5py>=3.7.0